OTR-457: Fix build
