+++ apps/ehr/src/features/visits/shared/components/patient/info/IdentifiersRow.tsx
@@ -5,7 +5,7 @@ import { Box, IconButton, Skeleton, Tooltip, Typography } from '@mui/material';
 import { Patient } from 'fhir/r4b';
 import { FC, ReactElement, useState } from 'react';
 import { dataTestIds } from 'src/constants/data-test-ids';
-import { getFriendlyPatientId } from 'utils';
+import { getFriendlyPatientId } from 'src/features/visits/shared/utils/friendly-patient-id.helper';
 
 type Props = {
   patient?: Patient;
