+++ packages/utils/lib/fhir/patient.ts
@@ -36,7 +36,6 @@ import {
   FHIR_IDENTIFIER_NPI,
   FHIR_IDENTIFIER_SYSTEM,
   filterResources,
-  FRIENDLY_PATIENT_ID_SYSTEM_BASE,
   getAllPractitionerCredentials,
   getCommunicationsAndSenders,
   getUniquePhonesNumbers,
@@ -835,14 +834,4 @@ export const mapGenderToLabel: { [name in Exclude<Patient['gender'], undefined>]
   female: 'Female',
   other: 'Intersex',
   unknown: 'Unknown',
-};
-
-export function getFriendlyPatientId(patient: Patient): string {
-  const system = getFriendlyPatientIdSystem();
-  return patient.identifier?.find((ident) => ident.system === system)?.value ?? '';
-}
-
-export const getFriendlyPatientIdSystem = (): string | undefined => {
-  const projectId = import.meta.env.VITE_APP_PROJECT_ID;
-  return projectId ? `${FRIENDLY_PATIENT_ID_SYSTEM_BASE}/${projectId}` : undefined;
 };
